Final Folder of the AEPS Residences
Final folder update where everything is organized and concentrated into a single folder for more efficient editing and use in the future. This folder is made from Visual Studio Code. Future comments will be added in RentalShop.cpp and AEPS Residences.cpp for future learning of the material after finals.